+++ src/parking_lot.rs
@@ -1,8 +1,10 @@
 use std::time::Instant;
 
 use parking_lot::{MutexGuard, RwLockReadGuard, RwLockWriteGuard};
+#[cfg(feature = "tracing")]
+use tracing::trace;
 
-use crate::lock_info::{GuardKind, Location, LockGuard, LockInfo, LockKind};
+use crate::lock_info::{call_location, GuardKind, Location, LockGuard, LockInfo, LockKind};
 
 #[derive(Debug)]
 pub struct Mutex<T> {
@@ -19,20 +21,41 @@ impl<T> Mutex<T> {
     }
 
     pub fn lock(&self) -> LockGuard<MutexGuard<'_, T>> {
+        let guard_kind = GuardKind::Lock;
+        let guard_location = call_location();
+        #[cfg(feature = "tracing")]
+        trace!("Acquiring a {:?} guard at {}", guard_kind, guard_location);
         let timestamp = Instant::now();
         let guard = self.lock.lock();
         let wait_time = timestamp.elapsed();
-        LockGuard::new(guard, GuardKind::Lock, &self.location, wait_time)
+        LockGuard::new(guard, guard_kind, &self.location, guard_location, wait_time)
     }
 
     pub fn try_lock(&self) -> Option<LockGuard<MutexGuard<'_, T>>> {
+        let guard_kind = GuardKind::Lock;
+        let guard_location = call_location();
+        #[cfg(feature = "tracing")]
+        trace!(
+            "Attempting to acquire a {:?} guard at {}",
+            guard_kind,
+            guard_location
+        );
         let timestamp = Instant::now();
-        let guard = self.lock.try_lock()?;
+        let guard = self.lock.try_lock().or_else(|| {
+            #[cfg(feature = "tracing")]
+            trace!(
+                "Failed to acquire a {:?} guard at {}",
+                guard_kind,
+                guard_location,
+            );
+            None
+        })?;
         let wait_time = timestamp.elapsed();
         Some(LockGuard::new(
             guard,
-            GuardKind::Lock,
+            guard_kind,
             &self.location,
+            guard_location,
             wait_time,
         ))
     }
@@ -62,17 +85,25 @@ impl<T> RwLock<T> {
     }
 
     pub fn read(&self) -> LockGuard<RwLockReadGuard<'_, T>> {
+        let guard_kind = GuardKind::Read;
+        let guard_location = call_location();
+        #[cfg(feature = "tracing")]
+        trace!("Acquiring a {:?} guard at {}", guard_kind, guard_location);
         let timestamp = Instant::now();
         let guard = self.lock.read();
         let wait_time = timestamp.elapsed();
-        LockGuard::new(guard, GuardKind::Read, &self.location, wait_time)
+        LockGuard::new(guard, guard_kind, &self.location, guard_location, wait_time)
     }
 
     pub fn write(&self) -> LockGuard<RwLockWriteGuard<'_, T>> {
+        let guard_kind = GuardKind::Write;
+        let guard_location = call_location();
+        #[cfg(feature = "tracing")]
+        trace!("Acquiring a {:?} guard at {}", guard_kind, guard_location);
         let timestamp = Instant::now();
         let guard = self.lock.write();
         let wait_time = timestamp.elapsed();
-        LockGuard::new(guard, GuardKind::Write, &self.location, wait_time)
+        LockGuard::new(guard, guard_kind, &self.location, guard_location, wait_time)
     }
 
     pub fn into_inner(self) -> T {